Update tax calculator.py
No changes, just wanna comment. I've written one part that will take a person's income and it's type and taxes it accordingly. Now that I'm written this, this function can probably be made cleaner by representing each of these if statements branches with a function. Or even lambda functions to make it more elegant. 

There is one problem. On line 30, if an exception has occured, and is eventually resolved, the previous iterations of the function throw errors. It will say that real_estate is not defined. However , this can be resolved by returning the function and ending it's execution. This discards the previous iterations of the function. But I want to deeply understand how this works.