Update tax calc.py
and with that I think I'm done. Learnt how to install libraries and deal with some of the potential errors that can come with installing. Utilised the PrettyTable Function. Using a module such as PrettyTable allowed me to create user-friendly output w/o having to deal with the process of creating a UI